extract models out of /agent dir
Summary:

Test Plan:
